requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
python-dotenv>=1.0.0
requests>=2.28.0
schedule>=1.2.0
psycopg2-binary>=2.9.0 